Module telegram_bot_raw::requests::_base

Structs

DetachedRequest

DetachedRequestType

Error
The Error type.

HttpRequest

HttpResponse

JsonIdResponse

JsonRequestType

JsonTrueToUnitResponse

Enums

Body

ErrorKind
The kind of an error.

Method

RequestUrl

Traits

JsonResponse

Request

RequestType

ResponseType

ToReplyRequest
Use this trait to convert a complex type to corresponding request and reply to the message.

ToRequest
Use this trait to convert a complex type to corresponding request and send it to the chat.