Module telegram_bot_raw::requests::send_chat_action

Structs

SendChatAction
Use this method when you need to tell the user that something is happening on the bot's side. The status is set for 5 seconds or less (when a message arrives from your bot, Telegram clients clear its typing status).

Enums

ChatAction
Strongly typed ChatAction. Instead of passing a String to the chat_action method, this is used.

Traits

CanSendChatAction
Send action to a chat.