Module telegram_bot_raw::requests::unban_chat_member

Structs

UnbanChatMember
Use this method to unban a previously kicked user in a supergroup or channel. The user will not return to the group or channel automatically, but will be able to join via link, etc. The bot must be an administrator in the group for this to work.

Traits

CanUnbanChatMemberForChat
Unban a previously kicked user in a supergroup or channel.

CanUnbanChatMemberForUser
Unban a previously kicked user in a supergroup or channel.