Module telegram_bot_raw::types::inline_query_result

Structs

InlineQueryResultArticle

InlineQueryResultAudio

InlineQueryResultCachedAudio

InlineQueryResultCachedDocument

InlineQueryResultCachedGif

InlineQueryResultCachedMpeg4Gif

InlineQueryResultCachedPhoto

InlineQueryResultCachedSticker

InlineQueryResultCachedVideo

InlineQueryResultCachedVoice

InlineQueryResultContact

InlineQueryResultDocument

InlineQueryResultGame

InlineQueryResultGif

InlineQueryResultLocation

InlineQueryResultMpeg4Gif

InlineQueryResultPhoto

InlineQueryResultVenue

InlineQueryResultVideo

InlineQueryResultVoice

InputContactMessageContent

InputLocationMessageContent

InputTextMessageContent

InputVenueMessageContent

Enums

InlineQueryResult
This object represents an incoming inline query. When the user sends an empty query, your bot could return some default or trending results.

InputMessageContent
This object represents the content of a message to be sent as a result of an inline query. Telegram clients currently support the following 4 types: