Module telegram_bot_raw::types::message

Structs

Audio
This object represents an audio file to be treated as music by the Telegram clients.

Contact
This object represents a phone contact.

Document
This object represents a general file (as opposed to photos, voice messages and audio files).

File
This object represents a file ready to be downloaded. The file can be downloaded via the link https://api.telegram.org/file/bot<token>/<file_path>. It is guaranteed that the link will be valid for at least 1 hour.

Forward
Information about the original message.

Location
This object represents a point on the map.

Message
This object represents a message.

MessageEntity
This object represents one special entity in a text message. For example, hashtags, usernames, URLs, etc.

PhotoSize
This object represents one size of a photo or a file / sticker thumbnail.

RawMessage
This object represents a message. Directly mapped.

RawMessageEntity
This object represents one special entity in a text message. For example, hashtags, usernames, URLs, etc. Directly mapped.

Sticker
This object represents a sticker.

UserProfilePhotos
This object represent a user's profile pictures.

Venue
This object represents a venue.

Video
This object represents a video file.

VideoNote
This object represents a video message (available in Telegram apps as of v.4.0).

Voice
This object represents a voice note.

Enums

ForwardFrom
Information about the source of the original message.

MessageEntityKind
Kind of the entity.

MessageKind
Kind of the message.