Struct telegram_bot_raw::types::message::ChannelPost

pub struct ChannelPost {
    pub id: MessageId,
    pub date: Integer,
    pub chat: Channel,
    pub forward: Option<Forward>,
    pub reply_to_message: Option<Box<MessageOrChannelPost>>,
    pub edit_date: Option<Integer>,
    pub kind: MessageKind,
}

This object represents a channel message.

Fields
id: MessageId

Unique message identifier inside this chat.

date: Integer

Date the message was sent in Unix time.

chat: Channel

Conversation the message belongs to.

forward: Option<Forward>

Information about the original message.

reply_to_message: Option<Box<MessageOrChannelPost>>

For replies, the original message. Note that the Message object in this field will not contain further reply_to_message fields even if it itself is a reply.

edit_date: Option<Integer>

Date the message was last edited in Unix time.

kind: MessageKind

Kind of the message.