Struct telegram_bot_raw::types::message::RawMessage

pub struct RawMessage {
    pub message_id: Integer,
    pub from: Option<User>,
    pub date: Integer,
    pub chat: Chat,
    pub forward_from: Option<User>,
    pub forward_from_chat: Option<Chat>,
    pub forward_from_message_id: Option<Integer>,
    pub forward_date: Option<Integer>,
    pub reply_to_message: Option<Box<Message>>,
    pub edit_date: Option<Integer>,
    pub text: Option<String>,
    pub entities: Option<Vec<MessageEntity>>,
    pub audio: Option<Audio>,
    pub document: Option<Document>,
    pub photo: Option<Vec<PhotoSize>>,
    pub sticker: Option<Sticker>,
    pub video: Option<Video>,
    pub voice: Option<Voice>,
    pub video_note: Option<VideoNote>,
    pub caption: Option<String>,
    pub contact: Option<Contact>,
    pub location: Option<Location>,
    pub venue: Option<Venue>,
    pub new_chat_members: Option<Vec<User>>,
    pub left_chat_member: Option<User>,
    pub new_chat_title: Option<String>,
    pub new_chat_photo: Option<PhotoSize>,
    pub delete_chat_photo: Option<True>,
    pub group_chat_created: Option<True>,
    pub supergroup_chat_created: Option<True>,
    pub channel_chat_created: Option<True>,
    pub migrate_to_chat_id: Option<Integer>,
    pub migrate_from_chat_id: Option<Integer>,
    pub pinned_message: Option<Box<Message>>,
}

This object represents a message. Directly mapped.

Fields
message_id: Integer

Unique message identifier inside this chat.

from: Option<User>

Sender, can be empty for messages sent to channels.

date: Integer

Date the message was sent in Unix time.

chat: Chat

Conversation the message belongs to.

forward_from: Option<User>

For forwarded messages, sender of the original message.

forward_from_chat: Option<Chat>

For messages forwarded from a channel, information about the original channel.

forward_from_message_id: Option<Integer>

For forwarded channel posts, identifier of the original message in the channel.

forward_date: Option<Integer>

For forwarded messages, date the original message was sent in Unix time.

reply_to_message: Option<Box<Message>>

For replies, the original message. Note that the Message object in this field will not contain further reply_to_message fields even if it itself is a reply.

edit_date: Option<Integer>

Date the message was last edited in Unix time.

text: Option<String>

For text messages, the actual UTF-8 text of the message, 0-4096 characters.

entities: Option<Vec<MessageEntity>>

For text messages, special entities like usernames, URLs, bot commands, etc. that appear in the text.

audio: Option<Audio>

Message is an audio file, information about the file.

document: Option<Document>

Message is a general file, information about the file.

photo: Option<Vec<PhotoSize>>

Message is a photo, available sizes of the photo.

sticker: Option<Sticker>

Message is a sticker, information about the sticker.

video: Option<Video>

Message is a video, information about the video.

voice: Option<Voice>

Message is a voice message, information about the file.

video_note: Option<VideoNote>

Message is a video note message, information about the file.

caption: Option<String>

Caption for the document, photo or video, 0-200 characters.

contact: Option<Contact>

Message is a shared contact, information about the contact.

location: Option<Location>

Message is a shared location, information about the location.

venue: Option<Venue>

Message is a venue, information about the venue.

new_chat_members: Option<Vec<User>>

New members that were added to the group or supergroup and information about them (the bot itself may be one of these members)

left_chat_member: Option<User>

A member was removed from the group, information about them (this member may be the bot itself)

new_chat_title: Option<String>

A chat title was changed to this value.

new_chat_photo: Option<PhotoSize>

A chat photo was change to this value.

delete_chat_photo: Option<True>

Service message: the chat photo was deleted.

group_chat_created: Option<True>

Service message: the group has been created.

supergroup_chat_created: Option<True>

Service message: the supergroup has been created. This field can‘t be received in a message coming through updates, because bot can’t be a member of a supergroup when it is created. It can only be found in reply_to_message if someone replies to a very first message in a directly created supergroup.

channel_chat_created: Option<True>

Service message: the channel has been created. This field can‘t be received in a message coming through updates, because bot can’t be a member of a channel when it is created. It can only be found in reply_to_message if someone replies to a very first message in a channel.

migrate_to_chat_id: Option<Integer>

The group has been migrated to a supergroup with the specified identifier.

migrate_from_chat_id: Option<Integer>

The supergroup has been migrated from a group with the specified identifier.

pinned_message: Option<Box<Message>>

Specified message was pinned. Note that the Message object in this field will not contain further reply_to_message fields even if it is itself a reply.