Module telegram_bot_raw::requests::_base

Structs

DetachedRequest

DetachedRequestType

Error

HttpRequest

HttpResponse

JsonIdResponse

JsonRequestType

JsonTrueToUnitResponse

MultipartRequestType

Enums

Body

Method

MultipartValue

RequestUrl

Traits

JsonResponse

Request

RequestType

ResponseType

ToMultipart

ToMultipartValue

ToReplyRequest
Use this trait to convert a complex type to corresponding request and reply to the message.

ToRequest
Use this trait to convert a complex type to corresponding request and send it to the chat.

Type Definitions

Multipart